syscalls/access02: Fix test with busybox
The problem here is that system executes '/bin/sh -c foo' then shell
tries to execute the filename but kernel does not recognize the
excutable format (of an empty file) hence libc fallbacks to
execvp('/bin/sh', argv[0], foo, ...) and because the filename in argv[0]
is passed as executable name it causes busybox return "applet does not
exist" error.

Hence we write "!/bin/sh" shebang to the executable file which tells the
shell that this is a shell script and we avoid running the execvp() in
the first place.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reported-by: Laurent GONZALEZ <ltp@gezedo.com>