Add read_all01 test
While using a shell script I wrote which dumps some of the contents of /sys
and /proc (submitted in a separate patch) I found some minor kernel
bugs. There is already a test specifically for /proc however it is attempting
to verify the behavior of particular files, where as this test is just doing
a basic sanity check and thus can be applied more generally.

Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>